+++ src/main/java/at/hannibal2/skyhanni/config/features/Garden.java
@@ -396,6 +396,15 @@ public class Garden {
     public boolean optimalSpeedWarning = false;
 
     @Expose
+    @ConfigOption(name = "Rancher Boots", desc = "Allows you to set the optimal speed in the rancher boots overlay by clicking on the presets.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 9)
+    public boolean optimalSpeedSignEnabled = true;
+
+    @Expose
+    public Position optimalSpeedSignPosition = new Position(200, 20, false, true);
+
+    @Expose
     @ConfigOption(name = "Custom Speed", desc = "Change the exact speed for every single crop.")
     @Accordion
     @ConfigAccordionId(id = 9)