+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 + Added farming weight next leaderboard position eta.
 + Added **Dicer Counter** - Count RNG drops for Melon Dicer and Pumpkin Dicer.
 + Added **Optimal Speed** - Show the optimal speed for your current tool in the hand. (Ty MelonKingDE for the values)
+  + Also available to select directly in the rancher boots overlay (contributed by nea)
 + Added **Warn When Close** - Warn with title and sound when the next crop milestone upgrade happens in 5 seconds. Useful for switching to a different pet for leveling.
 + Added **Money per Hour** - Displays the money per hour YOU get with YOUR crop/minute value when selling the items to bazaar.
 + Added farming contest timer.